+++ test/fixtures/valid/array-in-arrays.js
@@ -22,6 +22,6 @@ var ast =
 module.exports = {
 	ast: ast,
 	options: {
-		verbose: true
+		loc: true
 	}
 };