+++ test/fixtures/valid/array-only.js
@@ -5,6 +5,6 @@ var array = types.createArray;
 module.exports = {
 	ast: array([], location(1, 1, 0, 1, 3, 2)),
 	options: {
-		verbose: true
+		loc: true
 	}
 };